createbooking

+++ HBS.HariBySilke_2021.WebApi/DTOs/TreatmentDto.cs
@@ -1,11 +1,9 @@
-using System;
-
 namespace HBS.HariBySilke_2021.WebApi.DTOs
 {
     public class TreatmentDto
     {
         public string TreatmentName { get; set; }
         public int Price { get; set; }
-        public int Duration { get; set; }
+        public double Duration { get; set; }
     }
 }
